+++ src/platform/windows.rs
@@ -171,7 +171,7 @@ pub trait WindowBuilderExtWindows {
     /// `COINIT_APARTMENTTHREADED`) on the same thread. Note that winit may still attempt to initialize
     /// COM API regardless of this option. Currently only fullscreen mode does that, but there may be more in the future.
     /// If you need COM API with `COINIT_MULTITHREADED` you must initialize it before calling any winit functions.
-    /// See https://docs.microsoft.com/en-us/windows/win32/api/objbase/nf-objbase-coinitialize#remarks for more information.
+    /// See <https://docs.microsoft.com/en-us/windows/win32/api/objbase/nf-objbase-coinitialize#remarks> for more information.
     fn with_drag_and_drop(self, flag: bool) -> WindowBuilder;
 
     /// Forces a theme or uses the system settings if `None` was provided.